set policy for role module

+++ modules/mam/Role/Http/Controllers/RoleController.php
@@ -18,6 +18,7 @@ class RoleController extends Controller
 
     public function index()
     {
+        $this->authorize('index',Role::class);
         $roles = $this->repository->getAllRoles();
         return view('Role::index',compact('roles'));
     }
@@ -27,6 +28,7 @@ class RoleController extends Controller
      */
     public function create(PermissionService $permissionService)
     {
+        $this->authorize('index',Role::class);
         $permissions = $permissionService->getAllPermissions();
         return view('Role::create',compact('permissions'));
     }
@@ -36,6 +38,7 @@ class RoleController extends Controller
      */
     public function store(RoleRequest $request)
     {
+        $this->authorize('index',Role::class);
         $this->repository->storeRole($request->only('name','permissions'));
         alert()->success('ساخت مقام', 'مقام با موفقیت ساخته شد');
         return to_route('roles.index');
@@ -46,6 +49,7 @@ class RoleController extends Controller
      */
     public function edit(int $id,PermissionService $permissionService)
     {
+        $this->authorize('index',Role::class);
         $permissions = $permissionService->getAllPermissions();
         $role = $this->repository->findById($id);
         return view('Role::edit',compact('permissions','role'));
@@ -56,6 +60,7 @@ class RoleController extends Controller
      */
     public function update(RoleRequest $request, int $id)
     {
+        $this->authorize('index',Role::class);
         $this->repository->updateRole($id,$request->only('name','permissions'));
         alert()->success('ویرایش مقام', 'مقام با موفقیت ویرایش شد');
         return to_route('roles.index');
@@ -66,6 +71,7 @@ class RoleController extends Controller
      */
     public function destroy(int $id)
     {
+        $this->authorize('index',Role::class);
         $this->repository->deleteRole($id);
         alert()->success('حذف مقام', 'مقام با موفقیت حذف شد');
         return to_route('roles.index');
